Additional Dataset Builder
Simulates a game of blackjack given a number of decks, players, and when to discard the remaining deck (penetration). Every round each player's hand is reported, win/loss/draw, when their "hit" lost (this minus one is optimal stay hand), and counts cards that have been viewed.

This dataset can be altered for a turn-by-turn training set, or stay as is for round-by-round.